+++ src/lpc.h
@@ -5,12 +5,12 @@
 
 class LPC{
 public:
-    double ABS2(double a,double b) {return sqrt(a*a+b*b);}
-    int CalPath(int frame_no, int frame_shift, double *Pitch);
-    void vowel_markPath(int *vowel_mark, int n, int frame_shift, double *Pitch);
-    void SpecEnv(double *coeff, double *SpecEnv_mag, double *SpecEnv_pha);
-	double From_Data(double *data, double *lpc, int n, int m);
-	void Predict(double *coeff, double *prime, int m, double *data, long n);
+    float ABS2(float a,float b) {return sqrt(a*a+b*b);}
+    int CalPath(int frame_no, int frame_shift, float *Pitch);
+    void vowel_markPath(int *vowel_mark, int n, int frame_shift, float *Pitch);
+    void SpecEnv(float *coeff, float *SpecEnv_mag, float *SpecEnv_pha);
+	float From_Data(float *data, float *lpc, int n, int m);
+	void Predict(float *coeff, float *prime, int m, float *data, long n);
 };
 
 #endif
